+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Auto-generated docs of all script commands in Herc's script_commands.txt
+- Variable autocomplete
 
 ### Changed
 - Recoded the script parser in Tree-Sitter instead of ANTLR, this improved code analysis a lot.